refactor: rename ipc module to store
