+++ app/src/main/java/com/niki914/nexus/agentic/mod/HookLocalSettings.kt
@@ -2,7 +2,7 @@ package com.niki914.nexus.agentic.mod
 
 import android.content.Context
 import com.niki914.nexus.xposed.api.util.ContextProvider
-import com.niki914.nexus.ipc.XIpcBridge
+import com.niki914.nexus.store.XIpcBridge
 
 object HookLocalSettings {
