+++ app/src/main/java/com/niki914/nexus/agentic/mod/XService.kt
@@ -2,10 +2,10 @@ package com.niki914.nexus.agentic.mod
 
 import android.content.Context
 import com.niki914.nexus.xposed.api.util.ContextProvider
-import com.niki914.nexus.ipc.HostApp
-import com.niki914.nexus.ipc.IpcReadResult
-import com.niki914.nexus.ipc.IpcWriteResult
-import com.niki914.nexus.ipc.XIpcBridge
+import com.niki914.nexus.store.HostApp
+import com.niki914.nexus.store.IpcReadResult
+import com.niki914.nexus.store.IpcWriteResult
+import com.niki914.nexus.store.XIpcBridge
 
 object XService {
