+++ app/src/main/java/com/niki914/nexus/agentic/runtime/client/AgentRuntimeClient.kt
@@ -13,7 +13,7 @@ import com.niki914.nexus.agentic.runtime.ipc.IAgentRuntimeService
 import com.niki914.nexus.agentic.runtime.ipc.IAgentStoreService
 import com.niki914.nexus.agentic.runtime.ipc.IRenderFrameCallback
 import com.niki914.nexus.agentic.runtime.ipc.RenderFrame
-import com.niki914.nexus.ipc.XIpcBridge
+import com.niki914.nexus.store.XIpcBridge
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
