+++ app/src/test/java/com/niki914/nexus/agentic/app/ui/nexus/model/HomeChatControllerTest.kt
@@ -12,7 +12,7 @@ import com.niki914.nexus.agentic.repo.AppStateSettings
 import com.niki914.nexus.agentic.repo.AppStateSettingsCodec
 import com.niki914.nexus.agentic.repo.DomainSettingsStore
 import com.niki914.nexus.agentic.repo.XRepo
-import com.niki914.nexus.ipc.store.StoreDescriptorRegistry
+import com.niki914.nexus.store.StoreDescriptorRegistry
 import com.niki914.s3ss10n.ChatTurn
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
