+++ app/src/test/java/com/niki914/nexus/agentic/repo/FakeDomainSettingsStore.kt
@@ -1,7 +1,7 @@
 package com.niki914.nexus.agentic.repo
 
 import android.content.Context
-import com.niki914.nexus.ipc.store.StoreDescriptorRegistry
+import com.niki914.nexus.store.StoreDescriptorRegistry
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonElement
