+++ settings.gradle.kts
@@ -34,5 +34,5 @@ include(":app")
 include(":xposed-api")
 include(":xposed-runtime")
 include(":composebase")
-include(":ipc")
+include(":store")
 include(":agent-runtime")
