+++ store/src/main/AndroidManifest.xml
