+++ store/src/main/res/values-en/strings.xml
