+++ store/src/test/java/com/niki914/nexus/store/IpcJsonMutatorTest.kt
@@ -1,4 +1,4 @@
-package com.niki914.nexus.ipc.store
+package com.niki914.nexus.store
 
 import org.json.JSONObject
 import org.junit.Assert.assertEquals
