+++ store/src/test/java/com/niki914/nexus/store/XIpcStoreRepositoryTest.kt
@@ -1,4 +1,4 @@
-package com.niki914.nexus.ipc.store
+package com.niki914.nexus.store
 
 import android.content.Context
 import android.content.ContextWrapper
